+++ tools/test-js-with-node
@@ -119,6 +119,7 @@ EXEMPT_FILES = make_set(
         "web/src/message_fetch.js",
         "web/src/message_list.js",
         "web/src/message_list_data.js",
+        "web/src/message_list_hover.js",
         "web/src/message_list_tooltips.js",
         "web/src/message_list_view.js",
         "web/src/message_lists.js",
